+++ src/App.php
@@ -69,7 +69,7 @@ class App
     /**
      * Get the session manager
      *
-     * @return Jaxon\Contracts\App\Session
+     * @return Jaxon\Contracts\Session
      */
     public function session()
     {
